Pull up removed as now down as a hardware mod Signal now counts as a low

+++ stackdaemon.py
@@ -7,25 +7,31 @@ DEFAULT_INTERVAL = 5
 DEFAULT_PULSE_WIDTH = 0.1
 DEFAULT_HBT_PIN = 5
 DEFAULT_SIG_PIN = 6
+TESTING = False
 
 class StackDaemon(object):
     def __init__(
             self,
+            testing=False,
             hbt_pin=DEFAULT_HBT_PIN,
             sig_pin=DEFAULT_SIG_PIN,
             interval=DEFAULT_INTERVAL,
             pulse_width=DEFAULT_PULSE_WIDTH):
+        global TESTING
+        TESTING = testing
         GPIO.setmode(GPIO.BCM)
         GPIO.setwarnings(False) #stop it displaying warnings aout channel already in use
         GPIO.setup(hbt_pin, GPIO.OUT)
-        GPIO.setup(sig_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
+        GPIO.setup(sig_pin, GPIO.IN, pull_up_down=GPIO.PUD_OFF)
         self._hbt_pin = hbt_pin
         self._sig_pin = sig_pin
         self._interval = interval
         self._pulse_width = pulse_width
 
     def run(self):
-        GPIO.add_event_detect(self._sig_pin, GPIO.RISING, callback=sig_recieved)
+        if TESTING:
+            print "TESTING MODE"
+        GPIO.add_event_detect(self._sig_pin, GPIO.FALLING, callback=sig_recieved)
         while True:
             self._send_heartbeat()
             sleep(self._interval)
@@ -38,13 +44,16 @@ class StackDaemon(object):
 
     def send_signal(self):
         GPIO.setup(self._sig_pin, GPIO.OUT)
-        GPIO.output(self._sig_pin, GPIO.HIGH)
+        GPIO.output(self._sig_pin, GPIO.LOW)
         sleep(self._pulse_width)
-        GPIO.output(self._hbt_pin, GPIO.LOW)
+        GPIO.output(self._hbt_pin, GPIO.HIGH)
         GPIO.setup(self._sig_pin, GPIO.IN)
 
-def sig_recieved(channel):
-    system("sudo shutdown -h now")
+def sig_recieved(channeli):
+    if not TESTING:
+        system("sudo shutdown -h now")
+    else:
+        print "signal recieved"
 
 if __name__ == "__main__":
     STACK = StackDaemon()
